+++ api/src/main/java/btrplace/model/constraint/Offline.java
@@ -28,17 +28,13 @@ import java.util.List;
 
 /**
  * A constraint to force a node at being offline.
- * <p/>
- * The restriction provided by the constraint is discrete.
- * however, if the node are already offline, then
- * its state will be unchanged.
  *
  * @author Fabien Hermenier
  */
 public class Offline extends SatConstraint {
 
     /**
-     * Instantiate constraints for a collection of nodes.
+     * Instantiate discrete constraints for a collection of nodes.
      *
      * @param nodes the nodes to integrate
      * @return the associated list of constraints
@@ -52,13 +48,24 @@ public class Offline extends SatConstraint {
     }
 
     /**
-     * Make a new constraint.
+     * Make a new discrete constraint.
      *
      * @param n the node to set offline
      */
     public Offline(Node n) {
-        super(Collections.<VM>emptyList(), Collections.singleton(n), false);
+        this(n, false);
     }
+
+    /**
+     * Make a new constraint.
+     *
+     * @param n          the node to set offline
+     * @param continuous {@code true} for a continuous restriction
+     */
+    public Offline(Node n, boolean continuous) {
+        super(Collections.<VM>emptyList(), Collections.singleton(n), continuous);
+    }
+
 
     @Override
     public SatConstraintChecker getChecker() {
@@ -66,12 +73,7 @@ public class Offline extends SatConstraint {
     }
 
     @Override
-    public boolean setContinuous(boolean b) {
-        return !b;
-    }
-
-    @Override
     public String toString() {
-        return "offline(nodes=" + getInvolvedNodes().iterator().next() + ", discrete)";
+        return "offline(nodes=" + getInvolvedNodes().iterator().next() + ", " + restrictionToString() + ")";
     }
 }