+++ api/src/main/java/btrplace/model/constraint/Running.java
@@ -28,17 +28,13 @@ import java.util.List;
 
 /**
  * A constraint to force a VM at being running.
- * <p/>
- * The restriction provided by the constraint is discrete.
- * however, if the VM is already running, then
- * its state will be unchanged.
  *
  * @author Fabien Hermenier
  */
 public class Running extends SatConstraint {
 
     /**
-     * Instantiate constraints for a collection of VMs.
+     * Instantiate discrete constraints for a collection of VMs.
      *
      * @param vms the VMs to integrate
      * @return the associated list of constraints
@@ -57,7 +53,17 @@ public class Running extends SatConstraint {
      * @param vm the VM to make running
      */
     public Running(VM vm) {
-        super(Collections.singleton(vm), Collections.<Node>emptySet(), false);
+        this(vm, false);
+    }
+
+    /**
+     * Make a new constraint.
+     *
+     * @param vm         the VM to make running
+     * @param continuous {@code true} for a continuous restriction
+     */
+    public Running(VM vm, boolean continuous) {
+        super(Collections.singleton(vm), Collections.<Node>emptySet(), continuous);
     }
 
     @Override
@@ -67,11 +73,6 @@ public class Running extends SatConstraint {
 
     @Override
     public String toString() {
-        return "running(vms=" + getInvolvedVMs().iterator().next() + ", discrete)";
-    }
-
-    @Override
-    public boolean setContinuous(boolean b) {
-        return !b;
+        return "running(vms=" + getInvolvedVMs().iterator().next() + ", " + restrictionToString() + ")";
     }
 }