+++ btrpsl/src/test/java/btrplace/btrpsl/constraint/OnlineBuilderTest.java
@@ -1,17 +1,18 @@
 /*
- * Copyright (c) 2013 University of Nice Sophia-Antipolis
+ * Copyright (c) 2014 University Nice Sophia Antipolis
  *
  * This file is part of btrplace.
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
  *
- * This program is distributed in the hope that it will be useful,
+ * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * You should have received a copy of the GNU General Public License
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
@@ -60,13 +61,13 @@ public class OnlineBuilderTest {
     @DataProvider(name = "goodOnlines")
     public Object[][] getGoodSignatures() {
         return new Object[][]{
-                new Object[]{"online(@N1);", 1},
-                new Object[]{">>online(@N[1..10]);", 10}
+                new Object[]{"online(@N1);", 1, true},
+                new Object[]{">>online(@N[1..10]);", 10, false}
         };
     }
 
     @Test(dataProvider = "goodOnlines")
-    public void testGoodSignatures(String str, int nbNodes) throws Exception {
+    public void testGoodSignatures(String str, int nbNodes, boolean c) throws Exception {
         ScriptBuilder b = new ScriptBuilder(new DefaultModel());
         Set<SatConstraint> cstrs = b.build("namespace test; VM[1..10] : tiny;\n@N[1..20] : defaultNode;\n" + str).getConstraints();
         Assert.assertEquals(cstrs.size(), nbNodes);
@@ -74,7 +75,7 @@ public class OnlineBuilderTest {
         for (SatConstraint x : cstrs) {
             Assert.assertTrue(nodes.addAll(x.getInvolvedNodes()));
             Assert.assertEquals(x.getInvolvedNodes().size(), 1);
-            Assert.assertEquals(x.isContinuous(), false);
+            Assert.assertEquals(x.isContinuous(), c);
         }
     }
 }