+++ btrpsl/src/test/java/btrplace/btrpsl/constraint/ReadyBuilderTest.java
@@ -1,17 +1,18 @@
 /*
- * Copyright (c) 2013 University of Nice Sophia-Antipolis
+ * Copyright (c) 2014 University Nice Sophia Antipolis
  *
  * This file is part of btrplace.
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
  *
- * This program is distributed in the hope that it will be useful,
+ * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- * You should have received a copy of the GNU General Public License
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
@@ -61,13 +62,13 @@ public class ReadyBuilderTest {
     @DataProvider(name = "goodReadys")
     public Object[][] getGoodSignatures() {
         return new Object[][]{
-                new Object[]{">>ready(VM1);", 1},
-                new Object[]{"ready(VM[1..10]);", 10}
+                new Object[]{">>ready(VM1);", 1, false},
+                new Object[]{"ready(VM[1..10]);", 10, true}
         };
     }
 
     @Test(dataProvider = "goodReadys")
-    public void testGoodSignatures(String str, int nbVMs) throws Exception {
+    public void testGoodSignatures(String str, int nbVMs, boolean c) throws Exception {
         ScriptBuilder b = new ScriptBuilder(new DefaultModel());
         Set<SatConstraint> cstrs = b.build("namespace test; VM[1..10] : tiny;\n@N[1..20] : defaultNode;" + str).getConstraints();
         Assert.assertEquals(cstrs.size(), nbVMs);
@@ -75,7 +76,7 @@ public class ReadyBuilderTest {
         for (SatConstraint x : cstrs) {
             Assert.assertTrue(x instanceof Ready);
             Assert.assertTrue(vms.addAll(x.getInvolvedVMs()));
-            Assert.assertEquals(x.isContinuous(), false);
+            Assert.assertEquals(x.isContinuous(), c);
         }
     }
 }