+++ CHANGES.md
@@ -3,6 +3,7 @@ Release notes
 
 version ?????????
 ----------------------
+- continuous or discrete restriction mode for state constraints, ban, fence
 - btrpsl joined the repository
 
 version 0.38.2 - 19 Jun 2014